Fix cast issue

+++ OpenSim/Services/InventoryService/XInventoryService.cs
@@ -312,7 +312,7 @@ namespace OpenSim.Services.InventoryService
             {
                 if (xFolder.version > check.Version)
                     return false;
-                check.Version = xFolder.version;
+                check.Version = (ushort)xFolder.version;
                 xFolder = ConvertFromOpenSim(check);
                 return m_Database.StoreFolder(xFolder);
             }
